Move database proxy manager service into company service
