+++ src/backend/app/Http/Controllers/CompanyController.php
@@ -14,7 +14,6 @@ use App\Services\UserService;
 use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Artisan;
-use MPM\DatabaseProxy\DatabaseProxyManagerService;
 
 class CompanyController extends Controller {
     public function __construct(
@@ -22,7 +21,6 @@ class CompanyController extends Controller {
         private CompanyDatabaseService $companyDatabaseService,
         private PluginsService $pluginsService,
         private UserService $userService,
-        private DatabaseProxyManagerService $databaseProxyManagerService,
         private MpmPluginService $mpmPluginService,
         private RegistrationTailService $registrationTailService,
         private MainSettingsService $mainSettingsService,
@@ -44,7 +42,7 @@ class CompanyController extends Controller {
         ]);
 
         // Create Admin user and DatabaseProxy
-        $this->databaseProxyManagerService->runForCompany(
+        $this->companyService->runForCompany(
             $company->getId(),
             fn () => $this->userService->create(
                 [
@@ -58,7 +56,7 @@ class CompanyController extends Controller {
         );
 
         // Set some meaningful settings by default
-        $this->databaseProxyManagerService->runForCompany(
+        $this->companyService->runForCompany(
             $company->getId(),
             function () use ($company, $usageType) {
                 $mainSettings = $this->mainSettingsService->getAll()->first();
@@ -70,7 +68,7 @@ class CompanyController extends Controller {
         );
 
         // Plugin and Registration Tail magic
-        return $this->databaseProxyManagerService->runForCompany(
+        return $this->companyService->runForCompany(
             $company->getId(),
             function () use ($company, $plugins) {
                 // Prompt new users to configure their default settings
@@ -107,8 +105,6 @@ class CompanyController extends Controller {
     }
 
     public function get($email): ApiResource {
-        $databaseProxy = $this->databaseProxyManagerService->findByEmail($email);
-
-        return ApiResource::make($this->companyService->getByDatabaseProxy($databaseProxy));
+        return ApiResource::make($this->companyService->findByEmail($email));
     }
 }
