+++ src/backend/app/modules/TenantResolver/ApiResolvers/DataExportResolver.php
@@ -3,11 +3,11 @@
 namespace MPM\TenantResolver\ApiResolvers;
 
 use App\Exceptions\ValidationException;
+use App\Services\CompanyService;
 use Illuminate\Http\Request;
-use MPM\DatabaseProxy\DatabaseProxyManagerService;
 
 class DataExportResolver implements ApiResolverInterface {
-    public function __construct(private DatabaseProxyManagerService $databaseProxyManager) {}
+    public function __construct(private CompanyService $companyService) {}
 
     public function resolveCompanyId(Request $request): int {
         $segments = $request->segments();
@@ -16,7 +16,7 @@ class DataExportResolver implements ApiResolverInterface {
         }
 
         $email = $segments[3];
-        $databaseProxy = $this->databaseProxyManager->findByEmail($email);
+        $databaseProxy = $this->companyService->findByEmail($email);
         $companyId = $databaseProxy->getCompanyId();
 
         return (int) $companyId;
