+++ src/backend/database/seeders/PluginsSeeder.php
@@ -7,13 +7,13 @@ use Database\Factories\CalinCredentialFactory;
 use Database\Factories\CsvDataFactory;
 use Database\Factories\WaveMoneyCredentialFactory;
 use Illuminate\Database\Seeder;
-use MPM\DatabaseProxy\DatabaseProxyManagerService;
+use App\Services\CompanyService;
 
 class PluginsSeeder extends Seeder {
     public function __construct(
-        private DatabaseProxyManagerService $databaseProxyManagerService,
+        private CompanyService $companyService,
     ) {
-        $this->databaseProxyManagerService->buildDatabaseConnectionDemoCompany();
+        $this->companyService->buildDatabaseConnectionDemoCompany();
     }
 
     /**
