+++ src/backend/database/seeders/TargetSeeder.php
@@ -5,14 +5,14 @@ namespace Database\Seeders;
 use App\Models\ConnectionGroup;
 use App\Models\SubTarget;
 use App\Models\Target;
+use App\Services\CompanyService;
 use Illuminate\Database\Seeder;
-use MPM\DatabaseProxy\DatabaseProxyManagerService;
 
 class TargetSeeder extends Seeder {
     public function __construct(
-        private DatabaseProxyManagerService $databaseProxyManagerService,
+        private CompanyService $companyService,
     ) {
-        $this->databaseProxyManagerService->buildDatabaseConnectionDemoCompany();
+        $this->companyService->buildDatabaseConnectionDemoCompany();
     }
 
     /**
