+++ src/backend/database/seeders/TenantSeeder.php
@@ -9,14 +9,12 @@ use App\Services\RegistrationTailService;
 use App\Services\UserService;
 use App\Utils\DemoCompany;
 use Illuminate\Database\Seeder;
-use MPM\DatabaseProxy\DatabaseProxyManagerService;
 
 class TenantSeeder extends Seeder {
     public function __construct(
         private CompanyDatabaseService $companyDatabaseService,
         private CompanyService $companyService,
         private UserService $userService,
-        private DatabaseProxyManagerService $databaseProxyManagerService,
         private RegistrationTailService $registrationTailService,
         private MainSettingsService $mainSettingsService,
     ) {}
@@ -43,7 +41,7 @@ class TenantSeeder extends Seeder {
         ]);
 
         // Create Admin user and DatabaseProxy
-        $this->databaseProxyManagerService->runForCompany(
+        $this->companyService->runForCompany(
             $company->getId(),
             fn () => $this->userService->create(
                 [
@@ -57,7 +55,7 @@ class TenantSeeder extends Seeder {
         );
 
         // Set some meaningful settings by default
-        $this->databaseProxyManagerService->runForCompany(
+        $this->companyService->runForCompany(
             $company->getId(),
             function () {
                 $mainSettings = $this->mainSettingsService->getAll()->first();
@@ -73,7 +71,7 @@ class TenantSeeder extends Seeder {
 
         // Plugin and Registration Tail magic
         // TBD: For now, only Registration Tail
-        $this->databaseProxyManagerService->runForCompany(
+        $this->companyService->runForCompany(
             $company->getId(),
             function () {
                 // Do not prompt demo users to configure their default settings
