+++ src/backend/database/seeders/TicketSeeder.php
@@ -8,6 +8,7 @@ use App\Models\MaintenanceUsers;
 use App\Models\MiniGrid;
 use App\Models\Person\Person;
 use App\Models\User;
+use App\Services\CompanyService;
 use Database\Factories\TicketFactory;
 use Database\Factories\TicketOutsourceFactory;
 use Database\Factories\TicketUserFactory;
@@ -15,13 +16,12 @@ use Illuminate\Console\View\Components\Info;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use Inensus\Ticket\Models\TicketCategory;
-use MPM\DatabaseProxy\DatabaseProxyManagerService;
 
 class TicketSeeder extends Seeder {
     public function __construct(
-        private DatabaseProxyManagerService $databaseProxyManagerService,
+        private CompanyService $companyService,
     ) {
-        $this->databaseProxyManagerService->buildDatabaseConnectionDemoCompany();
+        $this->companyService->buildDatabaseConnectionDemoCompany();
     }
 
     private $amount = 100;
