+++ src/backend/packages/inensus/viber-messaging/src/Http/Controllers/ViberCredentialController.php
@@ -2,17 +2,17 @@
 
 namespace Inensus\ViberMessaging\Http\Controllers;
 
+use App\Services\CompanyService;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\URL;
 use Inensus\ViberMessaging\Http\Requests\ViberCredentialRequest;
 use Inensus\ViberMessaging\Http\Resources\ViberResource;
 use Inensus\ViberMessaging\Services\ViberCredentialService;
-use MPM\DatabaseProxy\DatabaseProxyManagerService;
 
 class ViberCredentialController extends Controller {
     public function __construct(
         private ViberCredentialService $credentialService,
-        private DatabaseProxyManagerService $databaseProxyManagerService,
+        private CompanyService $companyService,
     ) {}
 
     public function show(): ViberResource {
@@ -20,7 +20,7 @@ class ViberCredentialController extends Controller {
     }
 
     public function update(ViberCredentialRequest $request): ViberResource {
-        $databaseProxy = $this->databaseProxyManagerService->findByEmail(auth('api')->user()->email);
+        $databaseProxy = $this->companyService->findByEmail(auth('api')->user()->email);
         $companyId = $databaseProxy->getCompanyId();
         $apiToken = $request->input('api_token');
         $id = $request->input('id');
